add a default decoder injector, move sts handler to pkg
