+++ pkg/admission/handlers/sts_test.go
@@ -1,4 +1,4 @@
-package admission
+package handlers
 
 import (
 	"context"
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/kanopy-platform/hedgetrimmer/internal/admission"
 	"github.com/kanopy-platform/hedgetrimmer/internal/limitrange"
 	"github.com/stretchr/testify/assert"
 	admissionv1 "k8s.io/api/admission/v1"
@@ -13,7 +14,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
+	kadmission "sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 )
 
 type MockMutator struct {
@@ -38,7 +39,7 @@ func TestSTSHandler(t *testing.T) {
 	mm := MockMutator{}
 
 	scheme := runtime.NewScheme()
-	decoder, err := admission.NewDecoder(scheme)
+	decoder, err := kadmission.NewDecoder(scheme)
 	assert.NoError(t, err)
 
 	handler := NewSTSHandler(&mm)
@@ -72,14 +73,6 @@ func TestSTSHandler(t *testing.T) {
 		mutated bool
 	}{
 		{
-			lrerr:  fmt.Errorf("Fail"),
-			reject: true,
-			msg:    "Deny for lr error",
-		},
-		{
-			msg: "Allow for a namespace with no limitranges",
-		},
-		{
 			reject: true,
 			config: &limitrange.Config{},
 			merr:   fmt.Errorf("Fail"),
@@ -98,13 +91,12 @@ func TestSTSHandler(t *testing.T) {
 	}
 
 	for _, test := range tests {
-		mlr.SetErr(test.lrerr)
 		mm.SetSpec(test.pts)
 		mm.SetErr(test.merr)
 
-		ctx := context.WithValue(ctx.Background(), LimitRangeContextTypeMemory, test.config)
+		ctx := context.WithValue(context.Background(), admission.LimitRangeContextTypeMemory, test.config)
 
-		resp := handler.Handle(ctx, admission.Request{AdmissionRequest: ar})
+		resp := handler.Handle(ctx, kadmission.Request{AdmissionRequest: ar})
 		assert.Equal(t, test.reject, !resp.Allowed, test.msg)
 		assert.Equal(t, test.mutated, len(resp.Patches) > 0)
 	}
